feat: 实现验证码服务及网关集成
- 新增验证码服务(VerifyServer)及相关配置
- 实现网关服务(GateServer)与验证码服务的gRPC通信
- 添加邮件发送功能用于验证码分发
- 更新CMake配置以支持gRPC和protobuf
- 重构.gitignore文件
- 修改错误码类型为int32以兼容gRPC

+++ src/GateServer/const.h
@@ -14,7 +14,7 @@ namespace http = beast::http;
 namespace net = boost::asio;
 using tcp = boost::asio::ip::tcp;
 
-enum class ErrorCodes : int16_t {
+enum class ErrorCodes : int32_t {
     SUCCESS = 0,
     ERROR_JSON = 1001,
     RPC_FAILED = 1002,
